# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top All Money API",
        "description": "Documentation de l'API top all money",
        "contact": {
            "email": "support@tonapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        },
        {
            "url": "https://api.tonapp.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/account-users": {
            "get": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "List account users",
                "operationId": "08041869d0e6616510081661f4889ad5",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isDirty",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "updatedAt:desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Create an account user",
                "operationId": "14a700a6ff633a7d5b75e7958481d4f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountUserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/account-users/{id}": {
            "get": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Show an account user",
                "operationId": "091697ae462a2870c85163176064c4a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Update an account user",
                "operationId": "75e67e9f41b8a2957403bd507b932669",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountUserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Soft delete",
                "operationId": "e43867d467e34dc8dc6f98c65a0ca9a4",
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Partially update an account user",
                "operationId": "246b1c537ea02151fd5c0826933d7064",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountUserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/account-users/bulk": {
            "post": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Bulk upsert account users",
                "operationId": "ba3cd8c66c38bda8d04bcb052d1a0169",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {}
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/account-users/import": {
            "post": {
                "tags": [
                    "AccountUsers"
                ],
                "summary": "Import account users from CSV",
                "operationId": "668892f78ff8da1b986d4d1d004544cb",
                "responses": {
                    "200": {
                        "description": "Imported"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AccountUserCreateRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "account": {
                        "type": "string",
                        "nullable": true
                    },
                    "user": {
                        "type": "string",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AccountUserUpdateRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "role": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}